create competitions app
