+++ rso_backend/api/serializers.py
@@ -16,8 +16,7 @@ from api.constants import (DOCUMENTS_RAW_EXISTS, EDUCATION_RAW_EXISTS,
                            PRIVACY_RAW_EXISTS, REGION_RAW_EXISTS,
                            STATEMENT_RAW_EXISTS, TOO_MANY_EDUCATIONS)
 from api.utils import create_first_or_exception, get_is_trusted
-from events.models import (Сompetition, СompetitionApplications,
-                           СompetitionParticipants, Event,
+from events.models import (Event,
                            EventAdditionalIssue, EventApplications,
                            EventDocument, EventDocumentData, EventIssueAnswer,
                            EventOrganizationData, EventParticipants,
@@ -39,6 +38,7 @@ from users.models import (MemberCert, RSOUser, UserDocuments, UserEducation,
                           UserPrivacySettings, UserProfessionalEducation,
                           UserRegion, UserStatementDocuments,
                           UserVerificationRequest)
+from competitions.models import CompetitionParticipants, CompetitionApplications, Competitions
 
 
 class PositionSerializer(serializers.ModelSerializer):
@@ -1615,25 +1615,25 @@ class DetachmentSerializer(BaseUnitSerializer):
         return data
 
     def get_status(self, obj):
-        if not СompetitionParticipants.objects.filter(
+        if not CompetitionParticipants.objects.filter(
             Q(detachment=obj) & Q(junior_detachment__isnull=False) |
             Q(detachment__isnull=False) & Q(junior_detachment=obj)
         ).exists():
             return None
-        if СompetitionParticipants.objects.filter(
+        if CompetitionParticipants.objects.filter(
             Q(detachment=obj) & Q(junior_detachment__isnull=False)
         ).exists():
             return 'Наставник'
         return 'Старт'
 
     def get_nomination(self, obj):
-        if not СompetitionParticipants.objects.filter(
+        if not CompetitionParticipants.objects.filter(
             Q(detachment=obj) & Q(junior_detachment__isnull=False) |
             Q(detachment__isnull=False) & Q(junior_detachment=obj) |
             Q(junior_detachment=obj)
         ).exists():
             return None
-        if СompetitionParticipants.objects.filter(
+        if CompetitionParticipants.objects.filter(
             Q(detachment=obj) & Q(junior_detachment__isnull=False) |
             Q(detachment__isnull=False) & Q(junior_detachment=obj)
         ).exists():
@@ -1641,7 +1641,7 @@ class DetachmentSerializer(BaseUnitSerializer):
         return 'Дебют'
 
     def get_tandem_partner(self, obj):
-        participants = СompetitionParticipants.objects.filter(
+        participants = CompetitionParticipants.objects.filter(
             Q(detachment=obj) | Q(junior_detachment=obj)
         ).first()
         if participants:
@@ -2205,7 +2205,7 @@ class ShortDetachmentCompititionSerializer(BaseShortUnitSerializer):
 
 class СompetitionSerializer(serializers.ModelSerializer):
     class Meta:
-        model = Сompetition
+        model = Competitions
         fields = '__all__'
 
 
@@ -2215,7 +2215,7 @@ class СompetitionApplicationsObjectSerializer(serializers.ModelSerializer):
     detachment = ShortDetachmentCompititionSerializer()
 
     class Meta:
-        model = СompetitionApplications
+        model = CompetitionApplications
         fields = (
             'id',
             'competition',
@@ -2228,7 +2228,7 @@ class СompetitionApplicationsObjectSerializer(serializers.ModelSerializer):
 
 class СompetitionApplicationsSerializer(serializers.ModelSerializer):
     class Meta:
-        model = СompetitionApplications
+        model = CompetitionApplications
         fields = (
             'id',
             'competition',
@@ -2246,8 +2246,8 @@ class СompetitionApplicationsSerializer(serializers.ModelSerializer):
 
     def validate(self, attrs):
         request = self.context.get('request')
-        applications = СompetitionApplications.objects.all()
-        participants = СompetitionParticipants.objects.all()
+        applications = CompetitionApplications.objects.all()
+        participants = CompetitionParticipants.objects.all()
         if request.method == 'POST':
             MIN_DATE = (f'{settings.DATE_JUNIOR_SQUAD[2]}'
                         f'.{settings.DATE_JUNIOR_SQUAD[1]}.'
@@ -2303,7 +2303,7 @@ class СompetitionParticipantsObjectSerializer(serializers.ModelSerializer):
     junior_detachment = ShortDetachmentCompititionSerializer()
 
     class Meta:
-        model = СompetitionParticipants
+        model = CompetitionParticipants
         fields = (
             'id',
             'competition',
@@ -2315,7 +2315,7 @@ class СompetitionParticipantsObjectSerializer(serializers.ModelSerializer):
 
 class СompetitionParticipantsSerializer(serializers.ModelSerializer):
     class Meta:
-        model = СompetitionParticipants
+        model = CompetitionParticipants
         fields = (
             'id',
             'competition',
