+++ rso_backend/api/tasks.py
@@ -1,7 +1,6 @@
 import logging
 
 from celery import shared_task
-from django.contrib.auth import get_user_model
 from rest_framework import status
 from rest_framework.response import Response
