+++ rso_backend/competitions/models.py
@@ -1,0 +1,131 @@
+from django.db import models
+
+
+class Competitions(models.Model):
+    name = models.CharField(
+        verbose_name='Название',
+        max_length=255,
+        unique=True
+    )
+    created_at = models.DateTimeField(
+        verbose_name='Дата и время создания',
+        auto_now_add=True
+    )
+    start_date = models.DateField(
+        verbose_name='Дата начала',
+        null=True,
+        blank=True
+    )
+    end_date = models.DateField(
+        verbose_name='Дата окончания',
+        null=True,
+        blank=True
+    )
+
+    def __str__(self):
+        return f'{self.name} id {self.id}'
+
+    class Meta:
+        verbose_name_plural = 'Конкурсы'
+        verbose_name = 'Конкурс'
+
+
+class CompetitionApplications(models.Model):
+    competition = models.ForeignKey(
+        to='Competitions',
+        on_delete=models.CASCADE,
+        related_name='competition_applications',
+        verbose_name='Конкурс',
+    )
+    detachment = models.ForeignKey(
+        to='headquarters.Detachment',
+        on_delete=models.CASCADE,
+        related_name='competition_applications',
+        verbose_name='Отряд',
+        null=True,
+        blank=True
+    )
+    junior_detachment = models.ForeignKey(
+        to='headquarters.Detachment',
+        on_delete=models.CASCADE,
+        related_name='junior_competition_applications',
+        verbose_name='Младший отряд',
+    )
+    created_at = models.DateTimeField(
+        verbose_name='Дата и время создания заявки',
+        auto_now_add=True
+    )
+    is_confirmed_by_junior = models.BooleanField(
+        verbose_name='Подтверждено младшим отрядом',
+        default=False
+    )
+
+    def __str__(self):
+        return (f'Заявка id {self.id} на участие '
+                f'в конкурсе id {self.competition.id}')
+
+    class Meta:
+        verbose_name_plural = 'Заявки на участие в конкурсе'
+        verbose_name = 'Заявка на участие в конкурсе'
+        constraints = [
+            models.UniqueConstraint(
+                fields=('competition', 'detachment'),
+                name='unique_detachment_application_for_competition'
+            ),
+            models.UniqueConstraint(
+                fields=('competition', 'junior_detachment'),
+                name='unique_junior_detachment_application'
+            ),
+            models.UniqueConstraint(
+                fields=('detachment', 'junior_detachment'),
+                name='unique_tandem_application'
+            )
+        ]
+
+
+class CompetitionParticipants(models.Model):
+    competition = models.ForeignKey(
+        to='Competitions',
+        on_delete=models.CASCADE,
+        related_name='competition_participants',
+        verbose_name='Конкурс',
+    )
+    junior_detachment = models.ForeignKey(
+        to='headquarters.Detachment',
+        on_delete=models.CASCADE,
+        related_name='junior_competition_participants',
+        verbose_name='Младший отряд',
+    )
+    detachment = models.ForeignKey(
+        to='headquarters.Detachment',
+        on_delete=models.CASCADE,
+        related_name='competition_participants',
+        verbose_name='Отряд',
+        null=True,
+        blank=True
+    )
+    created_at = models.DateTimeField(
+        verbose_name='Дата и время создания заявки',
+        auto_now_add=True
+    )
+
+    def __str__(self):
+        return f'Участник(и) id {self.id} в конкурсе id {self.competition.id}'
+
+    class Meta:
+        verbose_name_plural = 'Участники конкурса'
+        verbose_name = 'Участник конкурса'
+        constraints = [
+            models.UniqueConstraint(
+                fields=('competition', 'detachment'),
+                name='unique_detachment_participant'
+            ),
+            models.UniqueConstraint(
+                fields=('competition', 'junior_detachment'),
+                name='unique_junior_detachment_participant'
+            ),
+            models.UniqueConstraint(
+                fields=('detachment', 'junior_detachment'),
+                name='unique_tandem_participant'
+            )
+        ]
