+++ rso_backend/events/admin.py
@@ -4,8 +4,7 @@ from events.models import (Event, EventAdditionalIssue, EventApplications,
                            EventDocument, EventDocumentData, EventIssueAnswer,
                            EventOrganizationData, EventParticipants,
                            EventTimeData, EventUserDocument,
-                           MultiEventApplication, Сompetition,
-                           СompetitionApplications, СompetitionParticipants)
+                           MultiEventApplication,)
 
 
 class EventTimeDataInline(admin.TabularInline):
@@ -60,6 +59,3 @@ admin.site.register(EventParticipants)
 admin.site.register(EventApplications)
 admin.site.register(EventIssueAnswer)
 admin.site.register(EventUserDocument)
-admin.site.register(Сompetition)
-admin.site.register(СompetitionParticipants)
-admin.site.register(СompetitionApplications)
