+++ rso_backend/rso_backend/settings.py
@@ -77,6 +77,7 @@ INSTALLED_APPS += [
     'users.apps.UsersConfig',
     'headquarters.apps.HeadquartersConfig',
     'events.apps.EventsConfig',
+    'competitions.apps.CompetitionsConfig',
 ]
 
 MIDDLEWARE = [
